requirements.txt
PyYAML>=6.0
beautifulsoup4>=4.12.0
